Joint Distributions for Stock Status Ratios

A function to generate joint distributions for stock status ratios (SSB/SSBref and F/Fref where ref can be MSY, SSB40, F40, etc.) using a Multivariate Log-Normal Distribution. The function produces a Kobe plot, maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values which can be input into SSplotEnsemble().

Usage

SSdeltaMVLN(
  ss3rep,
  Fref = NULL,
  years = NULL,
  mc = 5000,
  weight = 1,
  run = "MVLN",
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE,
  variance_method = "ww2019",
  bias_correct_mean = FALSE
)

Arguments

ss3rep

from r4ss::SS_output

Fref

Choice of Fratio c("MSY","Btgt"), correponding to F_MSY and F_Btgt

years

single year or vector of years for mvln

mc

number of monte-carlo simulations

weight

weighting option for model ensembles weight*mc

run

qualifier for model run

plot

option to show plot

addprj

include forecast years

ymax

ylim maximum

xmax

xlim maximum

legendcex

Allows to adjust legend cex

verbose

Report progress to R GUI?

variance_method

Specify method for approximating the variance and covariance of the multivariate lognormal distribution

ww2019

Use the approximation from Walter and Winker 2019 (Default for backwards compatibility).

2T

Use the delta-method and a 2nd order Taylor series approximation.

bias_correct_mean

Specify if the Stock Synthesis MLE estimates should be bias corrected when used to specify the mean of the multivariate lognormal distribution

FALSE

Do not apply bias correction (Default for backwards compatibility).

TRUE

Apply bias correction.

Value

output list of maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values, and kobe plot

Author

Henning Winker (JRC-EC)

Nicholas Ducharme-Barth (PIFSC)