Joint Distributions for Stock Status Ratios

A function to generate kobe pdfs from a Multivariate Log-Normal Distribution including plotting option. This version is used for older versions of Stock Synthesis. For newer versions (v3.24 and newer), use SSdeltaMVLN().

SSdeltaMVLN_old(
  ss3rep,
  status = c("Bratio", "F"),
  quants = c("SSB", "Recr"),
  Fref = c("MSY", "Ftrg"),
  years = NULL,
  mc = 5000,
  weight = 1,
  run = "MVLN",
  plot = TRUE,
  addtrj = TRUE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)

Arguments

ss3rep

from r4ss::SSgetoutput()$replist1

status

covarying stock status quantaties to extract from Hessian

quants

additional stock quantaties to extract from Hessian

Fref

Choice of reference point for stock SSB/XFref=c("MSY","Ftrg"),only if F_report_basis: 0 or 3

years

single year or vector of years for mvln

mc

number of monte-carlo simulations

weight

weighting option for model ensembles weight*mc

run

qualifier for model run

plot

option to show plot

addtrj

TRUE/FALSE add trajectory of years. Default is TRUE

ymax

ylim maximum

xmax

xlim maximum

legendcex

Allows to adjust legend cex

verbose

Report progress to R GUI?

Value

output list of kobe objects and mle's

Author

Henning Winker (JRC-EC)