MCMC Joint Distributions

A function to generate joint distributions for stock status ratios (SSB/SSBref and F/Fref where ref can be MSY, SSB40, F40, etc.) using MCMC. The function produces a Kobe plot, maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values which can be input into SSplotEnsemble(). Use for Stock Synthesis models older than 3.24, for later versions use SSdiagsMCMC().

SSdiagsMCMC_ancient(
  mcmc,
  ss3rep,
  Fref = NULL,
  years = NULL,
  run = "MCMC",
  thin = 1,
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)

Arguments

mcmc

file path for folder with the derived_posteriors.sso file

ss3rep

from r4ss::SS_output

Fref

Choice of Fratio c("MSY","Btgt), correponding to F_MSY and F_Btgt

years

single year or vector of years for mvln

run

qualifier for model run

thin

option to use additional thinning

plot

option to show plot

addprj

include forecast years

ymax

ylim maximum

xmax

xlim maximum

legendcex

Allows to adjust legend cex

verbose

Report progress to R GUI?

Value

output list maximum likelihood estimates and the MCMC posterier distributions of the Kobe values and a Kobe plot

Author

Henning Winker (JRC-EC), Massimiliano and Laurence Kell (Sea++)