Hindcasting Cross-Validations

Plots one-step ahead hindcasting cross-validations and computes MASE from prediction redisuals

SSplotEnsemble(
  kb,
  summaryoutput,
  subplots = c("stock", "harvest", "SSB", "F", "Recr", "Catch"),
  models = "all",
  quantiles = c(0.025, 0.975),
  ylabs = NULL,
  endyrvec = "default",
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = FALSE,
  use_pdf = FALSE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "l",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = NULL,
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  legendindex = NULL
)

Arguments

kb

SSdeltaMVLN $kb type output

summaryoutput

List created by r4ss::SSummarize(). TODO: Verify

subplots

option to "Bratio","Fvalue","SSB", "F", "Recr","Catch"

models

option to manually subset the models in kb$run

quantiles

quantiles for uncertainty in plots. Default is (.025,.075)

ylabs

yaxis labels for quants final year of values to show for each model. By default it is set to the

endyrvec

ending year specified in each model.

plot

Deprecated. Plots (and subplots) are drawn to the active plot device by default (TRUE), and the option to disable this, via FALSE, is unused.

print

Deprecated. Please use 'print_plot'.

print_plot

print to PNG files?

png

Deprecated. Please use 'use_png'.

use_png

Draw plots in PNG format

pdf

Deprecated. Please use 'use_pdf'.

use_pdf

option for pdf plots (not tested for TRUE)

col

Optional vector of colors to be used for lines. Input NULL

pch

Optional vector of plot character values

lty

Optional vector of line types

lwd

Optional vector of line widths

tickEndYr

TRUE/FALSE switch to turn on/off extra axis mark at final year in timeseries plots.

xlim

Optional, years to use for x-axis. Default NULL uses all years available.

ylimAdj

Multiplier for ylim parameter. Allows additional white space

xaxs

Choice of xaxs parameter (see ?par for more info)

yaxs

Choice of yaxs parameter (see ?par for more info)

xylabs

TRUE or FALSE, include x- and y-axis labels. Defaults to TRUE

type

Type parameter passed to points (default 'o' overplots points on top of lines)

uncertainty

TRUE/FALSE include uncertainty intervals around SSB or F estimated timeseries. Defaults to TRUE.

legend

Add a legend to plot. TRUE by default.

legendlabels

Optional vector of labels to include in legend.

legendloc

Location of legend. Either a string like "topleft" or a vector of two numeric values representing the fraction of the maximum in the x and y dimensions, respectively. See ?legend for more info on the string options.

legendorder

Optional vector of model numbers that can be used to have the legend display the model names in an order that is different than that which is represented in the summary input object.

legendncol

Number of columns for the legend.

legendcex

Allows to adjust legend cex

legendsp

Space between legend labels

pwidth

Width of plot

pheight

Height of plot

punits

Units for PNG file

res

Resolution for PNG file

ptsize

Point size for PNG file

cex.main

Character expansion for plot titles

plotdir

Directory where PNG or PDF files will be written. By default it will be the directory where the model was run.

filenameprefix

Additional text to append to PNG or PDF file names. It will be separated from default name by an underscore.

par

list of graphics parameter values passed to par() function

verbose

Report progress to R GUI?

shadecol

uncertainty shading of hcxval horizon

shadealpha

Transparency adjustment used to make default shadecol

new

Deprecated. New plot windows are created by default (TRUE), and the option to disable this, via FALSE, is unused.

add

surpresses par() to create multiplot figs

mcmcVec

mcmc vector TODO TODO. Default is FALSE

indexQlabel

TRUE/FALSE include labels for indices. Default is TRUE (currently not used)

indexQdigits

Number of significant digits for catchability in legend. Default is 4

legendindex

Allows to add lengend for selected indices (plots)

Author

Mostly adopted from r4ss::SSplotComparisons by Taylor et al

Examples

if (FALSE) {

mvln <- SSdeltaMVLN(ss3sma, run = "SMA")
sspar(mfrow = c(3, 2), plot.cex = 0.7)
SSplotEnsemble(mvln$kb, ylabs = mvln$labels, add = T, verbose = F)
}