Uncertainty parameters for Stock Synthesis Generic plots.

Generalized version of ss3diags SSplots functions to specify uncertainty. Not intended for production.

Usage

SSplotGenericUncertainty(uncertainty = TRUE, mcmcVec = FALSE, models = "all")

Arguments

uncertainty

Logical flag to enable plots with uncertainty intervals. Either a single TRUE/FALSE value, or a vector of TRUE/FALSE values around SSB or F estimated timeseries, or a set of integers corresponding to the choice of models.

mcmcVec

Logical vector of TRUE/FALSE values (or single value) indicating whether input values are from MCMC or to use normal distribution around MLE.

models

Optional subset of the models of summaryoutput (or a similar field with a different name): a list created by the function r4ss::SSsummarize. Either "all" or a vector of numbers indicating columns in summary tables.