Residual plot

plots residuals for all indices as boxplot (color coded by fleet) with a loess showing systematic trends over time. This function is from the package JABBA.

Usage

SSplotJABBAres(
  ss3rep = ss3diags::simple,
  subplots = c("cpue", "len", "age", "size", "con")[1],
  seas = NULL,
  plot = TRUE,
  print = lifecycle::deprecated(),
  print_plot = FALSE,
  png = lifecycle::deprecated(),
  use_png = print_plot,
  pdf = lifecycle::deprecated(),
  use_pdf = FALSE,
  indexselect = NULL,
  miny = 3,
  col = NULL,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  xlim = "default",
  ylim = "default",
  ylimAdj = 1.1,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendlabels = "default",
  legendloc = "bottomleft",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  boxcol = grey(0.8, 0.5),
  new = TRUE,
  add = FALSE
)

Arguments

ss3rep

Stock Synthesis output as read by r4SS function SS_output

subplots

string of type of data to plot, 'cpue' for index of abundance data, 'len' or 'age' for length- or age-composition, 'size' for generalized size composition, or 'con' for conditional age-at-length data.

seas

string indicating how to treat data from multiple seasons 'comb' - combine seasonal data for each year and plot against Yr 'sep' - treat season separately, plotting against Yr.S. If is.null(seas), it is assumed that there is only one season and option 'comb' is used.

plot

Deprecated. Plots (and subplots) are drawn to the active plot device by default (TRUE), and the option to disable this, via FALSE, is unused.

print

Deprecated. Please use 'print_plot'.

print_plot

Option to print to PNG files

png

Deprecated. please use 'use_png'.

use_png

Draw plots in PNG format

pdf

Deprecated. Please use 'use_pdf'.

use_pdf

option for pdf plots (not tested for TRUE)

indexselect

Vector of fleet numbers for each model for which to compare

miny

minimum abs values of ylim

col

Optional vector of colors to be used for lines. Input NULL

pch

Optional vector of plot character values

lty

Optional vector of line types

lwd

Optional vector of line widths

tickEndYr

TRUE/FALSE switch to turn on/off extra axis mark at final year in time series plots.

xlim

Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data. (currently not used)

ylim

Optional, min and max values for the ylim to override the "default" value (-0.7, 0.5)

ylimAdj

Multiplier for ylim parameter. Allows additional white space

xaxs

Choice of x-axis parameter (see ?par for more info)

yaxs

Choice of y-axis parameter (see ?par for more info)

xylabs

TRUE or FALSE, include x- and y-axis labels

type

Type parameter passed to points (default 'o' overplots points on top of lines)

legend

Option to add a legend. TRUE by default.

legendlabels

Optional vector of labels to include in legend.

legendloc

Location of legend. Either a string like "topleft" or a vector of two numeric values representing the fraction of the maximum in the x and y dimensions, respectively. See ?legend for more info on the string options.

legendorder

Optional vector of model numbers that can be used to have the legend display the model names in an order that is different than that which is represented in the summary input object.

legendncol

Number of columns for the legend.

legendcex

Allows to adjust legend cex. Defaults to 1.

legendsp

Space between legend labels

legendindex

Allows to add legend for selected indices (plots)

pwidth

Width of plot

pheight

Height of plot

punits

Units for PNG file

res

Resolution for PNG file

ptsize

Point size for PNG file

cex.main

Character expansion for plot titles

plotdir

Directory where PNG or PDF files will be written. By default it will be the directory where the model was run.

filenameprefix

Additional text to append to PNG or PDF file names. It will be separated from default name by an underscore.

par

list of graphics parameter values passed to par() function

verbose

Report progress to R GUI?

boxcol

color boxes

new

Deprecated. New plot windows are created by default (TRUE), and the option to disable this, via FALSE, is unused.

add

supresses par() to create multiplot figs

Author

Henning Winker (JRC-EC)