KOBE phase plot

Generates a Kobe plot or phase plot illustrating the stock status uncertainty over SSB/SSBmsy and F/Fmsy

SSplotKobe(
  kb,
  joint = TRUE,
  year = NULL,
  posterior = c("points", "kernel"),
  xlab = expression(SSB/SSB[MSY]),
  ylab = expression(F/F[MSY]),
  ylim = NULL,
  xlim = NULL,
  fill = TRUE,
  showMedian = TRUE,
  legend = TRUE,
  legendpos = "right",
  legendcex = 0.7,
  legendruns = TRUE,
  yr.label = TRUE,
  yr.int = 5,
  verbose = TRUE
)

Arguments

kb

output from SSdeltaMVLN()$kb

joint

option FALSE shows individual runs

year

option to choose year for kobe, last year is default

posterior

visualization of posterior c("points","kernel")

xlab

graphic parameter

ylab

graphic parameter

ylim

graphic parameter

xlim

graphic parameter

fill

shows color-coded quadrants

showMedian

option for median trajectory. TRUE by default

legend

option for legend. TRUE by default.

legendpos

legend position

legendcex

legend size

legendruns

show legend for run labels

yr.label

show year along trajectory

yr.int

year intervals along trajectory

verbose

Output to R console. Default is TRUE

Value

Kobe Quadrant percentages