Retrospective-Forecast with One-Step Ahead Hindcasting

Plots retrospective pattern, including (optional) one-step ahead forecast and computes Mohn's Rho

Usage

SSplotRetro(
  summaryoutput,
  subplots = c("SSB", "F"),
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xlim = NULL,
  xmin = NULL,
  labels = NULL,
  ylim = NULL,
  forecast = TRUE,
  forecastrho = TRUE,
  showrho = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.7,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = grey(0.4, 0.6),
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  shadecol1 = grey(0.5, 0.4),
  indexQlabel = TRUE,
  indexQdigits = 4,
  shadealpha = 0.3
)

Arguments

summaryoutput

List created by r4ss::SSsummarize()

subplots

Optional vector of subplots to be created, "SSB", "Bratio", "Fvalue", "Recruits", "Index"

plot

Option to draw subplots and plot in the interface. Deprecated. Option to disable will be removed in future version.

print

print to PNG files? Deprecated. Please use print_plot.

print_plot

Option to print to PNG files

png

Deprecated, please use 'use_png'

use_png

Draw plots in PNG format

pdf

PDF plots. Deprecated. Please use use_pdf.

use_pdf

option for pdf plots (currently does not work when subplots specified)

models

Optional subset of the models described in r4ss function summaryoutput(). Either "all" or a vector of numbers indicating columns in summary tables.

endyrvec

Optional single year or vector of years representing the final year of values to show for each model. By default it is set to the ending year specified in each model.

xlim

optional xlim, which overwrites xmin

xmin

optional minimum year shown in plot (default first yr)

labels

yaxis lable for biomass (bony fish and sharks)

ylim

option to specify ylim range

forecast

if true one-step ahead forecasts are shown in plot

forecastrho

if true one-step ahead forecast rho value is denoted in plot

showrho

TRUE/FALSE include Mohn's rho value? Defaults to TRUE

col

Optional vector of colors to be used for lines. Input NULL

pch

Optional vector of plot character values

lty

Optional vector of line types

lwd

Optional vector of line widths

tickEndYr

TRUE/FALSE switch to turn on/off extra axis mark at final year in timeseries plots.

ylimAdj

Multiplier for ylim parameter. Allows additional white space

xaxs

Choice of xaxs parameter (see ?par for more info)

yaxs

Choice of yaxs parameter (see ?par for more info)

xylabs

TRUE or FALSE, include x- and y-axis labels. Defaults to TRUE

type

Type parameter passed to points (default 'o' overplots points on top of lines)

uncertainty

TRUE/FALSE include uncertainty intervals around SSB or F estimated timeseries. Defaults to TRUE.

legend

Add a legend?

legendlabels

Optional vector of labels to include in legend.

legendloc

Location of legend. Either a string like "topleft" or a vector of two numeric values representing the fraction of the maximum in the x and y dimensions, respectively. See ?legend for more info on the string options.

legendorder

Optional vector of model numbers that can be used to have the legend display the model names in an order that is different than that which is represented in the summary input object.

legendncol

Number of columns for the legend.

legendcex

Allows to adjust legend cex

legendsp

Space between legend labels

legendindex

Allows to add lengend for selected indices (plots)

pwidth

Width of plot

pheight

Height of plot

punits

Units for PNG file

res

Resolution for PNG file

ptsize

Point size for PNG file

cex.main

Character expansion for plot titles

plotdir

Directory where PNG or PDF files will be written. By default it will be the directory where the model was run.

filenameprefix

Additional text to append to PNG or PDF file names. It will be separated from default name by an underscore.

par

list of graphics parameter values passed to par() function

verbose

Report progress to R GUI?

shadecol

uncertainty shading of hcxval horizon

new

Deprecated. New plot windows are created by default (TRUE), and the option to disable this, via FALSE, is unused.

add

surpresses par() to create multiplot figs

mcmcVec

NOT TESTED Vector of TRUE/FALSE values (or single value) indicating mcmc values are used

shadecol1

uncertainty shading of early years not affected by hindcast

indexQlabel

TRUE/FALSE, if TRUE add catchability to legend in plot of index fits (currently not used)

indexQdigits

Number of significant digits for catchability in legend

shadealpha

set the transparency level (alpha) of the area of uncertainty. Defalut to 0.3 (currently not used)

Author

Henning Winker (JRC-EC) and Laurance Kell (Sea++)