Residual Diagnostics

Function for residual diagnostics. Plots residuals and 3x sigma limits for indices or mean age or length and outputs a runs test table. Note, if you do not want to plot the residuals, use function ss3diags::SSrunstest.

Usage

SSplotRunstest(
  ss3rep = ss3diags::simple,
  mixing = "less",
  subplots = c("cpue", "len", "age", "size", "con")[1],
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  indexselect = NULL,
  miny = 1,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = "default",
  ylim = "default",
  ylimAdj = 1.4,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendloc = "top",
  legendcex = 1,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  new = TRUE,
  add = FALSE
)

Arguments

ss3rep

Stock Synthesis output as read by r4SS function SS_output

mixing

c("less","greater","two.sided"). Default less is checking for positive autocorrelation only

subplots

optional 'cpue' for index data, 'len' for length composition data, 'size' for generalized size composition data, 'age' for age composition data, or 'con' for conditional age at length data

plot

Deprecated. Plots (and subplots) are drawn to the active plot device by default (TRUE), the option to disable this via FALSE, is unused here.

print

Deprecated. Please use 'print_plot'.

print_plot

Option to print to PNG files

png

Deprecated, please use 'use_png'.

use_png

Draw plots in PNG format

pdf

PDF plots. Deprecated. Please use use_pdf.

use_pdf

option for pdf plots

indexselect

Vector of fleet numbers for each model for which to compare

miny

the absolute value of the min and max value for ylim, default is 1

pch

Optional vector of plot character values

lty

Optional vector of line types

lwd

Optional vector of line widths

tickEndYr

TRUE/FALSE switch to turn on/off extra axis mark at final year in timeseries plots.

xlim

Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data.

ylim

Optional, values for y-axis range to display on plot. Default = "default" will range from -1 to 1.

ylimAdj

Multiplier for ylim parameter. Allows additional white space

xaxs

Choice of xaxs parameter (see ?par for more info)

yaxs

Choice of yaxs parameter (see ?par for more info)

xylabs

TRUE or FALSE, include x- and y-axis labels

type

Type parameter passed to points (default 'o' overplots points on top of lines)

legend

Option to add a legend. TRUE by default.

legendloc

Location of legend. Either a string like "topleft" or a vector

legendcex

Allows to adjust legend cex

pwidth

Width of plot

pheight

Height of plot

punits

Units for PNG file

res

Resolution for PNG file

ptsize

Point size for PNG file

cex.main

Character expansion for plot titles

plotdir

Directory where PNG or PDF files will be written. By default it will be the directory where the model was run.

filenameprefix

Additional text to append to PNG or PDF file names. It will be separated from default name by an underscore.

par

list of graphics parameter values passed to par() function

verbose

TRUE or FALSE, should the progress be reported to R GUI?

new

Create new empty plot window (TRUE or FALSE)

add

suppresses par() to create multiplot figs

Value

a dataframe with runs test p-value, if the test has passed or failed, 3x sigma high and low limits, and the type of data used. Rows are for each fleet. Note, runs test passed if p-value > 0.05 (residuals are random) and failed if p-value < 0.5 (residuals are not random)

Author

Henning Winker (JRC-EC) and Laurance Kell (Sea++)