Hindcasting Cross-Validations of Multiple Models

Plots one-step ahead hindcasting cross-validations and computes MASE from prediction residuals. MASE is calculated as the average ratio of mean absolute error (MAE) of prediction residuals (MAE.PR) and naive predictions (MAE.base) MASE.adj sets the MAE.base to a minimum MAE.base.adj (default=0.1) MASE.adj allow passing (MASE<1) if MAE.PE < 0.1 and thus accurate if obs show very little annual variation

SSplotHCxval(
  retroSummary,
  subplots = c("cpue", "len", "age"),
  Season = "default",
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xmin = NULL,
  indexselect = NULL,
  MAE.base.adj = 0.1,
  show.mase.adj = TRUE,
  indexUncertainty = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  xlim = "default",
  ylimAdj = 1.15,
  ylim = NULL,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = grey(0.5, 0.4),
  shadecol2 = grey(0.5, 0.4),
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  indexfleets = 1,
  plot = TRUE,
  shadecol1 = grey(0.5, 0.4)
)

Arguments

retroSummary

List created by r4ss::SSsummarize() or ss3diags::SSretroComps()

subplots

optional use of c("cpue","len","age"), yet to be tested for age.

Season

option to specify Season - Default uses first available, i.e. usual Seas = 1

print

Deprecated. Please use 'print_plot'.

print_plot

Option to print to PNG files

png

Deprecated, please use 'use_png'.

use_png

Draw plots in PNG format

pdf

PDF plots. Deprecated. Please use use_pdf.

use_pdf

option for pdf plots

models

Optional subset of the models described in r4ss function summaryoutput(). Either "all" or a vector of numbers indicating columns in summary tables.

endyrvec

Optional single year or vector of years representing the final year of values to show for each model. By default it is set to the ending year specified in each model.

xmin

optional number first year shown in plot (if available)

indexselect

= Vector of fleet numbers for each model for which to compare

MAE.base.adj

minimum MASE denominator (naive predictions) for MASE.adj (default = 0.1)

show.mase.adj

if TRUE it show mase.adj in () in plot

indexUncertainty

Show fixed uncertainty intervals on index (not estimated)

col

Optional vector of colors to be used for lines. Input NULL

pch

Optional vector of plot character values

lty

Optional vector of line types

lwd

Optional vector of line widths

tickEndYr

TRUE/FALSE switch to turn on/off extra axis mark at final year in timeseries plots.

xlim

Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data. (currently not in use)

ylimAdj

Multiplier for ylim parameter. Allows additional white space

ylim

will over-write ylimAdj if specified

xaxs

Choice of xaxs parameter (see ?par for more info)

yaxs

Choice of yaxs parameter (see ?par for more info)

xylabs

TRUE or FALSE, include x- and y-axis labels

type

Type parameter passed to points (default 'o' overplots points on top of lines)

uncertainty

TRUE/FALSE include uncertainty intervals around SSB or F estimated timeseries. Defaults to TRUE.

legend

Option to add a legend. TRUE by default.

legendlabels

Optional vector of labels to include in legend.

legendloc

Location of legend. Either a string like "topleft" or a vector of two numeric values representing the fraction of the maximum in the x and y dimensions, respectively. See ?legend for more info on the string options.

legendorder

Optional vector of model numbers that can be used to have the legend display the model names in an order that is different than that which is represented in the summary input object.

legendncol

Number of columns for the legend.

legendcex

Allows to adjust legend cex

legendsp

Space between legend labels

legendindex

Allows to add legend for selected indices (plots)

pwidth

Width of plot

pheight

Height of plot

punits

Units for PNG file

res

Resolution for PNG file

ptsize

Point size for PNG file

cex.main

Character expansion for plot titles

plotdir

Directory where PNG or PDF files will be written. By default it will be the directory where the model was run.

filenameprefix

Additional text to append to PNG or PDF file names. It will be separated from default name by an underscore.

par

list of graphics parameter values passed to par() function

verbose

Report progress to R GUI?

shadecol

uncertainty shading of hcxval horizon

shadecol2

color for uncertainty in early years not affected by hindcast

shadealpha

Transparancy adjustment used to make default shadecol. (currently not used)

new

Create new empty plot window

add

suppresses par() to create multiplot figs

mcmcVec

NOT TESTED Vector of TRUE/FALSE values (or single value) indicating

indexQlabel

Add catchability to legend in plot of index fits (TRUE/FALSE)?

indexQdigits

Number of significant digits for catchability in legend

indexfleets

CHECK IF NEEDED or how to adjust indexfleets

plot

Deprecated. Plots (and subplots) are drawn to the active plot device by default (TRUE), and the option to disable this, via FALSE, is unused.

shadecol1

uncertainty shading of early years not affected by hindcast (currently not used)

Author

Henning Winker (JRC-EC) and Laurence Kell (Sea++)